+++ front-end/app/(auth)/landing.tsx
@@ -17,7 +17,7 @@ export default function Landing() {
     const { error } = await supabase.auth.signInWithPassword({ email, password });
     setLoading(false);
     if (error) Alert.alert("Login error", error.message);
-    else router.replace("/(tabs)");
+    else router.replace("/home");
   };
 
   const handleSignup = async () => {
@@ -27,6 +27,7 @@ export default function Landing() {
     setLoading(false);
     if (error) Alert.alert("Sign up error", error.message);
     else Alert.alert("Success", "Check your email!", [{ text: "OK", onPress: () => setActiveSheet(null) }]);
+    // note: email confirmation flow; user will land on / after they confirm
   };
 
   return (
