--- front-end/app/(tabs)/explore.tsx
@@ -1,12 +1,0 @@
-import { Image } from 'expo-image';
-import { Platform, StyleSheet } from 'react-native';
-import { View, Text } from 'react-native';
-
-export default function TabTwoScreen() {
-  return (
-    <View>
-      <Text>Hi</Text>
-    </View>
-  );
-}
-
