+++ front-end/app/(tabs)/profile.tsx
@@ -4,16 +4,12 @@ import { View,Text } from 'react-native';
 import PostForm from '@/components/PostForm';
 
 export default function HomeScreen() {
-  function handlePost(text:string){
-    console.log("New post:", text);
-  }
+ 
   return (
    <View className="flex-1 items-center justify-center bg-white">
       <Text className="text-xl font-bold text-purple-500">
-        Welcome to PURPLE!
+        profile      
       </Text>
-   
-      <PostForm onSubmit={handlePost} />
     </View>
   );
 }
